requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
google-api-python-client==2.142.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
python-dotenv==1.0.1


